requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-postgres>=0.0.6
langgraph>=0.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Caching
redis>=4.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Embeddings
sentence-transformers>=2.2.0

# Email functionality
reportlab>=4.0.0

# Web interface
streamlit>=1.28.0

# Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment and configuration
python-dotenv>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: Jupyter notebook support
jupyter>=1.0.0
ipython>=8.0.0
